backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
dnspython==2.3.0
requests==2.31.0
beautifulsoup4==4.12.2
python-whois==0.8.0
pandas==2.0.2
xlsxwriter==3.1.0
aiohttp==3.8.5
pydantic==1.10.8
pydantic[email]
openpyxl==3.1.2 